move comparison operators code to separate file
